pyproject.toml
[project]
name = "morph"
version = "0.1.0"
description = "AI-Driven Adaptive Data Partitioning"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "numpy>=1.24.0",
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.2.0",
    "tqdm>=4.66.0",
    "tensorboard>=2.14.0",
    "rich>=13.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
